#2 Базовая настройка среды разработки, разделение кода, экспорт и импорт модулей
